Add linting

+++ src/streamer/test_pattern_source.cpp
@@ -13,20 +13,34 @@ namespace {
 
 // 5-pixel-wide font for digits 0-9 and 'M'. Each entry is 7 rows of 5 bits.
 // Bit (1 << (4-x)) set = pixel on.
-struct Glyph { std::uint8_t rows[7]; };
+struct Glyph {
+    std::uint8_t rows[7];
+};
 
 // fmt-check: unused-function, keep it even if compiler thinks otherwise.
 static const Glyph kFont[] = {
-    ['0' - '0'] = {{0b01110, 0b10001, 0b10011, 0b10101, 0b11001, 0b10001, 0b01110}},
-    ['1' - '0'] = {{0b00100, 0b01100, 0b00100, 0b00100, 0b00100, 0b00100, 0b01110}},
-    ['2' - '0'] = {{0b01110, 0b10001, 0b00001, 0b00110, 0b01000, 0b10000, 0b11111}},
-    ['3' - '0'] = {{0b11111, 0b00010, 0b00100, 0b00010, 0b00001, 0b10001, 0b01110}},
-    ['4' - '0'] = {{0b00010, 0b00110, 0b01010, 0b10010, 0b11111, 0b00010, 0b00010}},
-    ['5' - '0'] = {{0b11111, 0b10000, 0b11110, 0b00001, 0b00001, 0b10001, 0b01110}},
-    ['6' - '0'] = {{0b00110, 0b01000, 0b10000, 0b11110, 0b10001, 0b10001, 0b01110}},
-    ['7' - '0'] = {{0b11111, 0b00001, 0b00010, 0b00100, 0b01000, 0b01000, 0b01000}},
-    ['8' - '0'] = {{0b01110, 0b10001, 0b10001, 0b01110, 0b10001, 0b10001, 0b01110}},
-    ['9' - '0'] = {{0b01110, 0b10001, 0b10001, 0b01111, 0b00001, 0b00010, 0b01100}},
+    // '0'-'0' keeps the index symmetric with the other digit entries below.
+    // NOLINTNEXTLINE(misc-redundant-expression)
+    ['0' -
+     '0'] = {{0b01110, 0b10001, 0b10011, 0b10101, 0b11001, 0b10001, 0b01110}},
+    ['1' - '0'] = {{0b00100, 0b01100, 0b00100, 0b00100, 0b00100, 0b00100,
+                    0b01110}},
+    ['2' - '0'] = {{0b01110, 0b10001, 0b00001, 0b00110, 0b01000, 0b10000,
+                    0b11111}},
+    ['3' - '0'] = {{0b11111, 0b00010, 0b00100, 0b00010, 0b00001, 0b10001,
+                    0b01110}},
+    ['4' - '0'] = {{0b00010, 0b00110, 0b01010, 0b10010, 0b11111, 0b00010,
+                    0b00010}},
+    ['5' - '0'] = {{0b11111, 0b10000, 0b11110, 0b00001, 0b00001, 0b10001,
+                    0b01110}},
+    ['6' - '0'] = {{0b00110, 0b01000, 0b10000, 0b11110, 0b10001, 0b10001,
+                    0b01110}},
+    ['7' - '0'] = {{0b11111, 0b00001, 0b00010, 0b00100, 0b01000, 0b01000,
+                    0b01000}},
+    ['8' - '0'] = {{0b01110, 0b10001, 0b10001, 0b01110, 0b10001, 0b10001,
+                    0b01110}},
+    ['9' - '0'] = {{0b01110, 0b10001, 0b10001, 0b01111, 0b00001, 0b00010,
+                    0b01100}},
 };
 
 void draw_label(std::uint8_t* base, int stride, int w, int h, int mon) {
@@ -34,9 +48,9 @@ void draw_label(std::uint8_t* base, int stride, int w, int h, int mon) {
     std::snprintf(text, sizeof(text), "M%d", mon);
 
     const int scale = std::max(4, std::min(w, h) / 80);  // pixel size per dot
-    const int gw = 5 * scale;      // glyph width
-    const int gh = 7 * scale;      // glyph height
-    const int gap = 2 * scale;     // gap between glyphs
+    const int gw = 5 * scale;                            // glyph width
+    const int gh = 7 * scale;                            // glyph height
+    const int gap = 2 * scale;                           // gap between glyphs
     const int margin = std::max(8, w / 40);
 
     int x0 = margin;
@@ -50,7 +64,10 @@ void draw_label(std::uint8_t* base, int stride, int w, int h, int mon) {
         for (int x = x0 - scale; x < x0 + total_w + scale && x < w; ++x) {
             if (x < 0 || y < 0) continue;
             auto* px = base + (std::size_t)y * stride + (std::size_t)x * 4;
-            px[0] = 0; px[1] = 0; px[2] = 0; px[3] = 255;
+            px[0] = 0;
+            px[1] = 0;
+            px[2] = 0;
+            px[3] = 255;
         }
     }
 
@@ -59,12 +76,16 @@ void draw_label(std::uint8_t* base, int stride, int w, int h, int mon) {
         const Glyph* g = nullptr;
         if (c == 'M') {
             // Custom M
-            static const Glyph M = {{0b10001, 0b11011, 0b10101, 0b10001, 0b10001, 0b10001, 0b10001}};
+            static const Glyph M = {{0b10001, 0b11011, 0b10101, 0b10001,
+                                     0b10001, 0b10001, 0b10001}};
             g = &M;
         } else if (c >= '0' && c <= '9') {
             g = &kFont[c - '0'];
         }
-        if (!g) { x0 += gw + gap; continue; }
+        if (!g) {
+            x0 += gw + gap;
+            continue;
+        }
         for (int gy = 0; gy < 7; ++gy) {
             for (int gx = 0; gx < 5; ++gx) {
                 if (!(g->rows[gy] & (1 << (4 - gx)))) continue;
@@ -72,11 +93,13 @@ void draw_label(std::uint8_t* base, int stride, int w, int h, int mon) {
                     for (int dx = 0; dx < scale; ++dx) {
                         int px_x = x0 + gx * scale + dx;
                         int px_y = y0 + gy * scale + dy;
-                        if (px_x < 0 || px_x >= w || px_y < 0 || px_y >= h) continue;
-                        auto* px = base + (std::size_t)px_y * stride + (std::size_t)px_x * 4;
-                        px[0] = 0;   // B
-                        px[1] = 255; // G
-                        px[2] = 255; // R - cyan
+                        if (px_x < 0 || px_x >= w || px_y < 0 || px_y >= h)
+                            continue;
+                        auto* px = base + (std::size_t)px_y * stride +
+                                   (std::size_t)px_x * 4;
+                        px[0] = 0;    // B
+                        px[1] = 255;  // G
+                        px[2] = 255;  // R - cyan
                         px[3] = 255;
                     }
                 }
@@ -99,8 +122,11 @@ void paint(AVFrame* f, std::int64_t idx, int mon) {
     const int scroll = static_cast<int>(idx * 4);
     const int box = 80;
     const int bx = (w - box) > 0 ? static_cast<int>((idx * 6) % (w - box)) : 0;
-    const int by = (h - box) > 0
-                       ? static_cast<int>((idx * 4) % (2 * (h - box))) : 0;
+    const int by =
+        (h - box) > 0
+            ? static_cast<int>((idx * 4) %
+                               (2 * static_cast<std::int64_t>(h - box)))
+            : 0;
     const int byy = by < (h - box) ? by : (2 * (h - box) - by);
 
     // Each monitor gets a different base hue.
@@ -114,12 +140,12 @@ void paint(AVFrame* f, std::int64_t idx, int mon) {
             // Blend the per-monitor hue into the color bar pattern.
             const int effbar = (bar + hue) % 6;
             switch (effbar) {
-                case 0: r = 255; break;
-                case 1: g = 255; break;
-                case 2: b = 255; break;
-                case 3: r = g = 255; break;
-                case 4: g = b = 255; break;
-                default: r = b = 255; break;
+            case 0: r = 255; break;
+            case 1: g = 255; break;
+            case 2: b = 255; break;
+            case 3: r = g = 255; break;
+            case 4: g = b = 255; break;
+            default: r = b = 255; break;
             }
             // Fade brightness with vertical position for a subtle gradient.
             const int shade = 60 + (y * 195) / (h ? h : 1);
@@ -183,8 +209,9 @@ int TestPatternSource::next_frame(AVFrame** out, std::int64_t& pts_usec) {
     const double fps =
         static_cast<double>(fmt_.fps_num) / (fmt_.fps_den ? fmt_.fps_den : 1);
     const auto target =
-        start_time_ + std::chrono::duration_cast<std::chrono::steady_clock::duration>(
-                          std::chrono::duration<double>(frame_index_ / fps));
+        start_time_ +
+        std::chrono::duration_cast<std::chrono::steady_clock::duration>(
+            std::chrono::duration<double>(frame_index_ / fps));
     std::this_thread::sleep_until(target);
 
     if (av_frame_make_writable(frame_) < 0) return -1;
